Created pre-build for GooCalCreds.xml creation.
Created a pre-build event for integration testing to generate a shell
xml file for the GooCalCreds.xml that is expected to exist on the dev
machine. Updated tests to validate the xml file is read and populates
the GooCalCreds properties with valid strings (not empty). Updated
integration tests to use new GooCalCreds.CreateAndValidate method when
instantiating a new GooCalCreds instance.